Replace VictoryGroup with refactored component
